Update with ops tests
