+++ docs/reference/utils/ops.md
@@ -14,10 +14,6 @@ keywords: Ultralytics YOLO, Utility Operations, segment2box, make_divisible, cli
 <br><br>
 
 ---
-## ::: ultralytics.utils.ops.coco80_to_coco91_class
-<br><br>
-
----
 ## ::: ultralytics.utils.ops.segment2box
 <br><br>
