+++ ultralytics/utils/files.py
@@ -143,13 +143,3 @@ def get_latest_run(search_dir='.'):
     """Return path to most recent 'last.pt' in /runs (i.e. to --resume from)."""
     last_list = glob.glob(f'{search_dir}/**/last*.pt', recursive=True)
     return max(last_list, key=os.path.getctime) if last_list else ''
-
-
-def make_dirs(dir='new_dir/'):
-    """Create directories."""
-    dir = Path(dir)
-    if dir.exists():
-        shutil.rmtree(dir)  # delete dir
-    for p in dir, dir / 'labels', dir / 'images':
-        p.mkdir(parents=True, exist_ok=True)  # make dir
-    return dir
